Moving sortObjectKeys to pipe with test coverage
