+++ src/app/common/pipes.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 
 import { AgoDatePipe } from './pipes/ago-date.pipe';
 import { KeysPipe } from './pipes/keys.pipe';
+import { SortObjectKeysPipe } from './pipes/sort-object-keys.pipe';
 import { UtcDatePipe } from './pipes/utc-date-pipe/utc-date.pipe';
 
 @NgModule({
@@ -9,11 +10,13 @@ import { UtcDatePipe } from './pipes/utc-date-pipe/utc-date.pipe';
 	exports: [
 		AgoDatePipe,
 		KeysPipe,
+		SortObjectKeysPipe,
 		UtcDatePipe
 	],
 	declarations: [
 		AgoDatePipe,
 		KeysPipe,
+		SortObjectKeysPipe,
 		UtcDatePipe
 	],
 	providers: []
@@ -22,4 +25,5 @@ export class PipesModule { }
 
 export { AgoDatePipe } from './pipes/ago-date.pipe';
 export { KeysPipe } from './pipes/keys.pipe';
+export { SortObjectKeysPipe } from './pipes/sort-object-keys.pipe';
 export { UtcDatePipe } from './pipes/utc-date-pipe/utc-date.pipe';
