+++ src/app/core/admin/audit/audit-view-change.component.html
@@ -17,11 +17,11 @@
 	<div class="row">
 		<div class="col-md-6">
 			<h4>Before</h4>
-			<pre>{{ sortObjectKeys(auditEntry?.audit?.object?.before) | json }}</pre>
+			<pre>{{ auditEntry?.audit?.object?.before | sortObjectKeys | json }}</pre>
 		</div>
 		<div class="col-md-6">
 			<h4>After</h4>
-			<pre>{{ sortObjectKeys(auditEntry?.audit?.object?.after) | json }}</pre>
+			<pre>{{ auditEntry?.audit?.object?.after | sortObjectKeys | json }}</pre>
 		</div>
 	</div>
 </div>
